DiscountedLineItemPortion as interface #1053
